refactor(acp): fold verified launch presets into driver
